Create rotateable module & incorporate it
